Reverted back to putting the grabbed objects into the RobotFilter.
I think this is the right way to handle grabbed bodies, because
further investigation has revealed that the erroneous collisions being
reported by the solver are *link* self-collisions where the grabbed
bodies collide with themselves.

This is because the KinBodyFilter is applied against all AffectedLinks,
which includes links of Grabbed bodies. Intuitively, this implies that
those links are expected to be in the same collision group as the other
robot links: i.e. RobotFilter.